put common functions into new file
